+++ src/main/java/org/folio/rest/impl/PutOrdersByIdHelper.java
@@ -1,115 +1,157 @@
 package org.folio.rest.impl;
 
+import static org.folio.orders.utils.HelperUtils.deletePoLine;
+import static org.folio.orders.utils.HelperUtils.getPoLines;
+import static org.folio.orders.utils.HelperUtils.operateOnSubObj;
+import static org.folio.orders.utils.SubObjects.PO_LINES;
+import static org.folio.orders.utils.SubObjects.PURCHASE_ORDER;
+import static org.folio.orders.utils.SubObjects.resourceByIdPath;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.CompletableFuture;
+
+import javax.ws.rs.core.Response;
+
+import org.apache.commons.lang.StringUtils;
+import org.folio.rest.jaxrs.model.CompositePurchaseOrder;
+import org.folio.rest.jaxrs.model.Error;
+import org.folio.rest.jaxrs.model.Errors;
+import org.folio.rest.jaxrs.model.PoLine;
+import org.folio.rest.jaxrs.resource.Orders.PutOrdersByIdResponse;
+
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Context;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
-import io.vertx.core.http.HttpHeaders;
+import io.vertx.core.http.HttpMethod;
+import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
-import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.LoggerFactory;
 import me.escoffier.vertx.completablefuture.VertxCompletableFuture;
-import org.folio.orders.rest.exceptions.HttpException;
-import org.folio.rest.jaxrs.model.CompositePurchaseOrder;
-import org.folio.rest.jaxrs.resource.Orders.PutOrdersByIdResponse;
-import org.folio.rest.tools.client.interfaces.HttpClientInterface;
 
-import javax.ws.rs.core.Response;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.concurrent.CompletableFuture;
+public class PutOrdersByIdHelper extends AbstractHelper {
 
-public class PutOrdersByIdHelper {
-  private static final Logger logger = LoggerFactory.getLogger(DeleteOrdersByIdHelper.class);
+  private final PostOrdersHelper postHelper;
+  private final PutOrderLineByIdHelper putLineHelper;
+  private final PostOrderLineHelper postOrderLineHelper;
 
-  private final HttpClientInterface httpClient;
-  private final Context ctx;
-  private final Handler<AsyncResult<javax.ws.rs.core.Response>> asyncResultHandler;
-  private final Map<String, String> okapiHeaders;
-
-  public PutOrdersByIdHelper(HttpClientInterface httpClient, Map<String, String> okapiHeaders,
-                             Handler<AsyncResult<Response>> asyncResultHandler, Context ctx) {
-    Map<String, String> customHeader = new HashMap<>();
-    customHeader.put(HttpHeaders.ACCEPT.toString(), "application/json, text/plain");
-    httpClient.setDefaultHeaders(customHeader);
-
-    this.httpClient = httpClient;
-    this.okapiHeaders = okapiHeaders;
-    this.ctx = ctx;
-    this.asyncResultHandler = asyncResultHandler;
+  public PutOrdersByIdHelper(Map<String, String> okapiHeaders,
+                             Handler<AsyncResult<Response>> asyncResultHandler, Context ctx, String lang) {
+    super(getHttpClient(okapiHeaders), okapiHeaders, asyncResultHandler, ctx, lang);
+    setDefaultHeaders(httpClient);
+    postHelper = new PostOrdersHelper(httpClient, okapiHeaders, asyncResultHandler, ctx, lang);
+    putLineHelper = new PutOrderLineByIdHelper(httpClient, okapiHeaders, asyncResultHandler, ctx, lang);
+    postOrderLineHelper = new PostOrderLineHelper(httpClient, okapiHeaders, asyncResultHandler, ctx, lang);
   }
 
-  /*
-   * Handle update order by doing Delete and Post
-   */
-  public CompletableFuture<Void> updateOrder(String id, String lang, CompositePurchaseOrder compPO,
-                                             Context vertxContext) {
+  public void updateOrderWithPoLines(String id, CompositePurchaseOrder compPO) {
+      getPoLines(id, lang, httpClient, ctx, okapiHeaders, logger)
+      .thenCompose(jsonObject -> {
+        JsonArray existedPoLinesArray = jsonObject.getJsonArray(PO_LINES);
+        return handlePoLines(compPO, existedPoLinesArray);
+      })
+      .thenCompose(aVoid -> updateOrder(id, compPO));
+  }
+
+  public CompletableFuture<Void> updateOrder(String id, CompositePurchaseOrder compPO) {
     CompletableFuture<Void> future = new VertxCompletableFuture<>(ctx);
-    GetOrdersByIdHelper getHelper = new GetOrdersByIdHelper(httpClient, okapiHeaders,
-      asyncResultHandler, vertxContext);
-    PostOrdersHelper postHelper = new PostOrdersHelper(httpClient, okapiHeaders, asyncResultHandler, vertxContext);
-    DeleteOrdersByIdHelper delHelper = new DeleteOrdersByIdHelper(httpClient, okapiHeaders, asyncResultHandler, vertxContext);
-    getHelper.getOrder(id, lang).thenAccept(existedCompPO ->
-      delHelper.deleteOrder(id, lang)
-        .thenRun(() -> {
-          compPO.setId(id);
-          compPO.setCreated(existedCompPO.getCreated());
-          compPO.setCreatedBy(existedCompPO.getCreatedBy());
-          postHelper.createPOandPOLines(compPO)
-            .thenAccept(withCompPO -> {
 
-              logger.info("Applying Funds...");
-              postHelper.applyFunds(withCompPO)
-                .thenAccept(withFunds -> {
+    logger.debug("Updating order...");
+    JsonObject purchaseOrder = convertToPurchaseOrder(compPO);
+    operateOnSubObj(HttpMethod.PUT, resourceByIdPath(PURCHASE_ORDER, id), purchaseOrder, httpClient, ctx, okapiHeaders, logger)
+      .thenAccept(v -> {
 
-                  logger.info("Updating Inventory...");
-                  postHelper.updateInventory(withFunds)
-                    .thenAccept(withInventory -> {
+        logger.info("Applying Funds...");
+        postHelper.applyFunds(compPO)
+          .thenAccept(withFunds -> {
 
-                      logger.info("Successfully Placed Order: " + JsonObject.mapFrom(withCompPO).encodePrettily());
-                      httpClient.closeClient();
-                      javax.ws.rs.core.Response response = PutOrdersByIdResponse.respond204();
-                      AsyncResult<javax.ws.rs.core.Response> result = Future.succeededFuture(response);
-                      asyncResultHandler.handle(result);
-                    })
-                    .exceptionally(this::handleError);
-                })
-                .exceptionally(this::handleError);
-            })
-            .exceptionally(this::handleError);
-        })
-        .exceptionally(this::handleError)
-    ).exceptionally(this::handleError);
+            logger.info("Updating Inventory...");
+            postHelper.updateInventory(withFunds)
+              .thenAccept(withInventory -> {
+
+                logger.info("Successfully Placed Order: " + JsonObject.mapFrom(compPO).encodePrettily());
+                httpClient.closeClient();
+                javax.ws.rs.core.Response response = PutOrdersByIdResponse.respond204();
+                AsyncResult<javax.ws.rs.core.Response> result = Future.succeededFuture(response);
+                asyncResultHandler.handle(result);
+              })
+              .exceptionally(this::handleError);
+          })
+          .exceptionally(this::handleError);
+      })
+      .exceptionally(this::handleError);
+
     return future;
   }
 
-  public Void handleError(Throwable throwable) {
-    final Future<javax.ws.rs.core.Response> result;
-
-    logger.error("Exception while updating orders", throwable.getCause());
-
-    final Throwable t = throwable.getCause();
-    if (t instanceof HttpException) {
-      final int code = ((HttpException) t).getCode();
-      final String message = ((HttpException) t).getMessage();
-      switch (code) {
-        case 404:
-          result = Future.succeededFuture(PutOrdersByIdResponse.respond404WithTextPlain(message));
-          break;
-        case 500:
-          result = Future.succeededFuture(PutOrdersByIdResponse.respond500WithTextPlain(message));
-          break;
-        default:
-          result = Future.succeededFuture(PutOrdersByIdResponse.respond500WithTextPlain(message));
-      }
+  private CompletableFuture<Void> handlePoLines(CompositePurchaseOrder compOrder, JsonArray poLinesFromStorage) {
+    List<CompletableFuture<?>> futures = new ArrayList<>();
+    if (poLinesFromStorage.isEmpty()) {
+      futures.addAll(processPoLinesCreation(compOrder, poLinesFromStorage));
     } else {
-      result = Future.succeededFuture(PutOrdersByIdResponse.respond500WithTextPlain(throwable.getMessage()));
+      futures.addAll(processPoLinesCreation(compOrder, poLinesFromStorage));
+      futures.addAll(processPoLinesUpdate(compOrder, poLinesFromStorage));
+      // The remaining unprocessed PoLines should be removed
+      poLinesFromStorage.stream().forEach(poLine -> futures.add(deletePoLine((JsonObject) poLine, httpClient, ctx, okapiHeaders, logger)));
     }
+    return VertxCompletableFuture.allOf(ctx, futures.toArray(new CompletableFuture[0]));
+  }
 
-    httpClient.closeClient();
+  private List<CompletableFuture<?>> processPoLinesUpdate(CompositePurchaseOrder compOrder, JsonArray poLinesFromStorage) {
+    List<CompletableFuture<?>> futures = new ArrayList<>();
+    for (int i = 0; i < poLinesFromStorage.size(); i++) {
+      JsonObject lineFromStorage = poLinesFromStorage.getJsonObject(i);
+      for (PoLine line : compOrder.getPoLines()) {
+        if (StringUtils.equals(lineFromStorage.getString("id"), line.getId())) {
+          futures.add(putLineHelper.updateOrderLine(line, lineFromStorage));
+          poLinesFromStorage.remove(i);
+          break;
+        }
+      }
+    }
+    return futures;
+  }
 
-    asyncResultHandler.handle(result);
+  private List<CompletableFuture<?>> processPoLinesCreation(CompositePurchaseOrder compOrder, JsonArray poLinesFromStorage) {
+    List<CompletableFuture<?>> futures = new ArrayList<>();
+    boolean updateInventory = compOrder.getWorkflowStatus() == CompositePurchaseOrder.WorkflowStatus.OPEN;
+    for (PoLine line : compOrder.getPoLines()) {
+      boolean isNew = true;
+      for (int i = 0; i < poLinesFromStorage.size(); i++) {
+        JsonObject lineFromStorage = poLinesFromStorage.getJsonObject(i);
+        if (StringUtils.equals(lineFromStorage.getString("id"), line.getId())) {
+          isNew = false;
+          break;
+        }
+      }
+      if (isNew) {
+        futures.add(postOrderLineHelper.createPoLine(line, updateInventory));
+      }
+    }
+    return futures;
+  }
 
-    return null;
+  @Override
+  Response buildErrorResponse(int code, Error error) {
+    final Response result;
+    switch (code) {
+      case 404:
+        result = PutOrdersByIdResponse.respond404WithTextPlain(error.getMessage());
+        break;
+      case 422:
+        result = PutOrdersByIdResponse.respond422WithApplicationJson(withErrors(error));
+        break;
+      default:
+        if (putLineHelper.getProcessingErrors().isEmpty()) {
+          result = PutOrdersByIdResponse.respond500WithTextPlain(error.getMessage());
+        } else {
+          Errors processingErrors = new Errors();
+          processingErrors.getErrors().addAll(putLineHelper.getProcessingErrors());
+          processingErrors.getErrors().add(error);
+          result = PutOrdersByIdResponse.respond500WithApplicationJson(processingErrors);
+        }
+    }
+    return result;
   }
 }